GUI/Watermarker Editor Image/requirements.txt
Pillow>=8.0